feat: add action badge
