+++ packages/app/src/systems/Chains/fuel/components/ActionBadge/ActionBadge.tsx
@@ -1,0 +1,19 @@
+import { cssObj } from '@fuel-ui/css';
+import { Badge } from '@fuel-ui/react';
+
+export const ActionBadge = () => {
+  return (
+    <Badge css={styles.actionBadge} intent="error">
+      Action Required
+    </Badge>
+  );
+};
+
+const styles = {
+  actionBadge: cssObj({
+    fontSize: '$xs',
+    lineHeight: 1,
+    padding: '$1',
+    textTransform: 'none',
+  }),
+};
