+++ packages/app/src/systems/Chains/fuel/components/ActionBadge/index.tsx
@@ -1,0 +1,1 @@
+export * from './ActionBadge';
