+++ packages/app/src/systems/Chains/fuel/components/index.tsx
@@ -1,0 +1,1 @@
+export * from './ActionBadge';
