+++ packages/app/src/systems/Chains/fuel/hooks/useTxFuelToEth.tsx
@@ -1,3 +1,4 @@
+import { cssObj } from '@fuel-ui/css';
 import { useTransaction } from '@fuels-portal/sdk-react';
 import { useInterpret, useSelector } from '@xstate/react';
 import type { ReceiptMessageOut } from 'fuels';
@@ -5,6 +6,7 @@ import { ReceiptType, fromTai64ToUnix } from 'fuels';
 import { useEffect, useMemo } from 'react';
 
 import { useEthAccountConnection } from '../../eth';
+import { ActionBadge } from '../components';
 import type { TxFuelToEthMachineState } from '../machines';
 import { txFuelToEthMachine } from '../machines';
 import { FUEL_UNITS } from '../utils';
@@ -51,7 +53,7 @@ const selectors = {
     const status = selectors.status(state);
 
     function getConfirmStatusText() {
-      if (status.isWaitingEthWalletApproval) return 'Action Required';
+      if (status.isWaitingEthWalletApproval) return <ActionBadge />;
       if (status.isConfirmTransactionDone) return 'Done!';
       return 'Action';
     }
