+++ packages/app/src/systems/Chains/fuel/index.tsx
@@ -1,4 +1,5 @@
 export * from './containers';
+export * from './components';
 export * from './hooks';
 export * from './machines';
 export * from './services';
